+++ web/documentserver-example/ruby/app/assets/javascripts/jscript.js
@@ -447,6 +447,7 @@ if (typeof jQuery != "undefined") {
 
     jq(document).on("click", ".try-editor", function (e) {
         var url = "sample?fileExt=" + e.target.attributes["data-type"].value;
+        url += "&lang=" + jq("#language").val();
         if (jq("#createSample").is(":checked")) {
             url += "&sample=true";
         }