+++ web/documentserver-example/ruby/app/controllers/home_controller.rb
@@ -88,7 +88,7 @@ class HomeController < ApplicationController
   def sample
     DocumentHelper.init(request.remote_ip, request.base_url)
     user = Users.get_user(params[:userId])
-    file_name = DocumentHelper.create_demo(params[:fileExt], params[:sample], user)
+    file_name = DocumentHelper.create_demo(params[:fileExt], params[:sample], params[:lang], user)
     redirect_to(controller: 'home', action: 'editor', fileName: file_name, userId: user.id, editorsMode: 'edit')
   end